+++ config/environment.js
@@ -1,0 +1,16 @@
+/* ==========================================================================
+   Settings for project environment. Used by JavaScript gulp build process.
+   ========================================================================== */
+
+'use strict';
+
+var paths = {
+  src:  './src',
+  dist: './dist',
+  lib:  './src/vendor',
+  test: './test'
+};
+
+module.exports = {
+  paths: paths
+}