+++ src/static/js/routes/careers/working-at-cfpb/index.js
@@ -4,18 +4,8 @@
 
 'use strict';
 
+var breakpointsConfig = require( '../../../config/breakpoints-config' );
 var MobileCarousel = require( '../../../modules/classes/MobileCarousel' );
 
-function init() {
-
-  // TODO: Remove this when per-page JS is introduced.
-  if ( document.querySelectorAll( '.careers-working-at-cfpb' ).length === 0 ) {
-    return;
-  }
-
-  // MobileCarousel with 599px breakpoint.
-  var mobileCarousel = new MobileCarousel( 599 );
-  mobileCarousel.enableOn( '.js-mobile-carousel' );
-}
-
-module.exports = { init: init };
+var mobileCarousel = new MobileCarousel( breakpointsConfig.mobile.max );
+mobileCarousel.enableOn( '.js-mobile-carousel' );